+++ redasm/formats/pe/pe_analyzer.cpp
@@ -9,7 +9,7 @@
 
 namespace REDasm {
 
-PEAnalyzer::PEAnalyzer(u32 petype, DisassemblerAPI *disassembler, const SignatureFiles& signatures): Analyzer(disassembler, signatures), m_petype(petype)
+PEAnalyzer::PEAnalyzer(u64 petype, DisassemblerAPI *disassembler, const SignatureFiles& signatures): Analyzer(disassembler, signatures), m_petype(petype)
 {
     ADD_WNDPROC_API(4, "DialogBoxA");
     ADD_WNDPROC_API(4, "DialogBoxW");
